map json file specs to Charliecloud arguments - issue
